Create README.md
Sharing economy platforms in the hospitality field, such as Airbnb, provide the user with helpful
information about available properties to stay in, such as their location, price, and host. However,
they are not providing information about the nearby places such as cafes, restaurants, activities and
transportation, or information about how likely the crime occurs in the property’s neighbourhood.
This information can effectively support the user’s decision of selecting a place to stay. Thus,
this project provides valuable information to explore property more in-depth. The Travel Guide
application lets the user know how often the crime occurs in a neighbourhood, information about
the property, its host, and summary of customers’ reviews . Moreover, this application lets users
know the nearby food venues, activities and transportation. In addition, the application provides
information about valuable locations like embassies. This application works with different data
sources, the Places API to provide information about the nearby places and transportation, DBPedia
API to provide information about activity venues, and the last source is the database, which contains
a large size of data that is combined from different sources. These sources are used to overcome the
information shortage in Airbnb and other platforms.